requirements.txt
fastapi>=0.108.0
uvicorn>=0.25.0
gunicorn>=20.1.0
pydantic
watchfiles
langsmith==0.1.93
langchain==0.2.14
mongoengine
langchain-community
pytest
langchain-groq
pytest-dotenv
sentry-sdk[fastapi]
requests
beautifulsoup4
urllib3
langchain-openai
langchain-chroma
langchain_text_splitters
scikit-learn
html2text
markdown
boto3
lxml
reportlab
Pillow